+++ ride_info/settings.py
@@ -27,6 +27,7 @@ INSTALLED_APPS = [
     "django_filters",
     "drf_spectacular",
     "api",
+    "rides",
 ]
 
 MIDDLEWARE = [
@@ -93,6 +94,12 @@ USE_TZ = True
 STATIC_URL = "static/"
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+
+AUTH_USER_MODEL = "rides.User"
+
+SIMPLE_JWT = {
+    "USER_ID_FIELD": "id_user",
+}
 
 REST_FRAMEWORK = {
     "DEFAULT_AUTHENTICATION_CLASSES": (